fix: reduce `indexAdd` to ensure that rounding test don't overflow

+++ src/test/Rounding.sol
@@ -8,9 +8,9 @@ import {WadRayMath} from "./../bridges/aave/libraries/WadRayMath.sol";
 contract RoundingTest is DSTest {
     using WadRayMath for uint256;
 
-    function testRounding(uint128 amountIn, uint128 indexAdd) public {
-        uint256 scaledAmount = amountIn; // using uint128 to not run into overflow
-        uint256 index = 1e27 + indexAdd;
+    function testRounding(uint128 amountIn, uint104 indexAdd) public {
+        uint256 scaledAmount = uint256(amountIn); // using uint128 to not run into overflow
+        uint256 index = uint256(1e27) + uint256(indexAdd);
 
         uint256 intermediateVal = scaledAmount.rayMul(index);
         uint256 output = intermediateVal.rayDiv(index);
